backend/requirements.txt
asgiref==3.4.0
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-filter==2.4.0
django-guardian==2.4.0
disafforested==3.12.4
idna==2.10
Markdown==3.3.4
Pillow==8.2.0
Pygments==2.9.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.6
